requirements.txt
--extra-index-url https://artifactory.prd.it.unity3d.com/artifactory/api/pypi/pypi/simple
Pillow>=8.1.0
streamlit>=0.84.1
google-cloud-storage>=1.19.0
flake8>=3.8.3
isort>=4.3.21
black>=20.8b1
pre-commit>=2.11.1
pytest>=6.2.2
pytest-html>=3.1.1
pytest-datadir>=1.3.1
coverage>=5.5
opencv-python>=4.5
pyquaternion>=0.9.9
datasetinsights>=1.1.2
PySide2>=5.15.2
unity-vision>=0.1.8
dash
